Moved CLI to seperate file
